Remove unused benchmarks and add benchmarks for emotion with glamorous style api
