--- packages/benchmarks/src/aphrodite.js
@@ -1,7 +1,0 @@
-import Box from './components/Box/aphrodite'
-import View from './components/View/aphrodite'
-
-export default {
-  Box,
-  View
-}
