--- packages/benchmarks/src/components/Box/new-css-in-js.js
@@ -1,31 +1,0 @@
-import styled from 'react-new-css-in-js'
-import View from '../View/new-css-in-js'
-
-const getColor = color => {
-  switch (color) {
-    case 0:
-      return '#222'
-    case 1:
-      return '#666'
-    case 2:
-      return '#999'
-    case 3:
-      return 'blue'
-    case 4:
-      return 'orange'
-    case 5:
-      return 'red'
-    default:
-      return 'transparent'
-  }
-}
-
-const Box = styled(View)`
-  flex-direction: ${props => (props.layout === 'column' ? 'column' : 'row')};
-  padding: ${props => (props.outer ? '4px' : '0')};
-  height: ${props => (props.fixed ? '20px' : 'auto')};
-  width: ${props => (props.fixed ? '20px' : 'auto')};
-  background-color: ${props => getColor(props.color)};
-`
-
-export default Box
