--- packages/benchmarks/src/components/Icons/styles.js
@@ -1,15 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-const styles = StyleSheet.create({
-  icon: {
-    display: 'inline-block',
-    fill: 'currentcolor',
-    height: '1.25em',
-    maxWidth: '100%',
-    position: 'relative',
-    userSelect: 'none',
-    verticalAlign: 'text-bottom'
-  }
-})
-
-export default styles
