+++ packages/benchmarks/src/components/View/emotion-css.js
@@ -5,7 +5,7 @@ import React from 'react'
 class View extends React.Component {
   render() {
     const { style, ...other } = this.props
-    return <div {...other} className={css([viewStyle, ...style])} />
+    return <div {...other} className={css(viewStyle, style)} />
   }
 }
