--- packages/benchmarks/src/components/View/new-css-in-js.js
@@ -1,19 +1,0 @@
-import styled from 'react-new-css-in-js'
-
-const View = styled('div')`
-  align-items: stretch;
-  border-width: 0;
-  border-style: solid;
-  box-sizing: border-box;
-  display: flex;
-  flex-basis: auto;
-  flex-direction: column;
-  flex-shrink: 0;
-  margin: 0;
-  padding: 0;
-  position: relative;
-  min-height: 0;
-  min-width: 0;
-`
-
-export default View
