--- packages/benchmarks/src/emotion-obj-style.js
@@ -1,7 +1,0 @@
-import Box from './components/Box/emotion-obj-style'
-import View from './components/View/emotion-obj-style'
-
-export default {
-  Box,
-  View
-}
