+++ packages/benchmarks/src/emotion-obj.js
@@ -1,0 +1,7 @@
+import Box from './components/Box/emotion-obj'
+import View from './components/View/emotion-obj'
+
+export default {
+  Box,
+  View
+}
