--- packages/benchmarks/src/jss.js
@@ -1,9 +1,0 @@
-import Box from './components/Box/jss'
-import View from './components/View/jss'
-
-const api = {
-  Box,
-  View
-}
-
-export default api
