--- packages/benchmarks/src/new-css-in-js.js
@@ -1,7 +1,0 @@
-import Box from './components/Box/new-css-in-js'
-import View from './components/View/new-css-in-js'
-
-export default {
-  Box,
-  View
-}
