--- packages/benchmarks/src/radium.js
@@ -1,9 +1,0 @@
-import Box from './components/Box/radium'
-import View from './components/View/radium'
-
-const api = {
-  Box,
-  View
-}
-
-export default api
