--- packages/benchmarks/src/react-native-stylesheet.js
@@ -1,9 +1,0 @@
-import Box from './components/Box/react-native-stylesheet'
-import View from './components/View/react-native-stylesheet'
-
-const api = {
-  Box,
-  View
-}
-
-export default api
