--- packages/benchmarks/src/react-native.js
@@ -1,9 +1,0 @@
-import Box from './components/Box/react-native'
-import Tweet from './components/Tweet'
-import { View } from 'react-native'
-
-export default {
-  Box,
-  Tweet,
-  View
-}
