--- packages/benchmarks/src/reactxp.js
@@ -1,7 +1,0 @@
-import Box from './components/Box/reactxp'
-import { View } from 'reactxp'
-
-export default {
-  Box,
-  View
-}
