--- packages/benchmarks/src/styled-components-primitives.js
@@ -1,7 +1,0 @@
-import Box from './components/Box/styled-components'
-import styled from 'styled-components/primitives'
-
-export default {
-  Box,
-  View: styled.View
-}
