--- packages/benchmarks/src/styletron.js
@@ -1,7 +1,0 @@
-import Box from './components/Box/styletron'
-import View from './components/View/styletron'
-
-export default {
-  Box,
-  View
-}
